[madepro] moved defaults as config submodule
