+++ madepro/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod config;
-pub mod defaults;
 pub mod models;
 pub mod solvers;
 pub mod utils;
